Add URI state machine with possible states
state.js works by calling the state() function, with a
state declaration inside of it.

Each of the properties in this object is going to be
either a function, or another state declaration object
wrapped in a call to state().

In this commit, we add 4 states that this object will
be able to find itself in.