Add implimentations of the .get and .set methods.
When you have a property that is only a function, it
will be the method implementation used when the
object is in that state, or any of it's child
substates.

In this case, a call to uri.get() will reach the
same function, no matter the state.

With an object in the 'Initial' state, a call
to uri.set(3000) will reach the method defined
in that state.