Conditionally change states based on input
You can access the state representation of any stateful object
using the this.state() accessor.

It has a number of methods that affect/read the object's state,
and is the lynchpin to the whole state.js library.

We are using one of the 3 mechanisms to change to another
state. They are:

1. `this.state().change('TargetState')`
2. `this.state().go('TargetState')`
3. `this.state('-> TargetState')`