Tests: Retrieving the currect state's name
You can use this.state().name to get the current name
of the state.

As you see here, it went to 'Initial' because we tagged
the state as that in the definition.

If we didn't do that, the state would have been '.', which
is where the get/listen and forward methods are defined.