Add connection mode and status machines.
To be able to manage the state of each service correctly, we are
also going to need to orchestrate whether we are forwarding/listening,
and whether the machine has been stopped or started.

We will coordinate all this using events, as each of the states,
and the machine as a whole acts as event emitters.