Introducing transition events and guards
I am now making use of the state machine's mechanisms to track
changes.

When you change state from one to the other, it will fire transition
events, that are also bound to functions on the instance.

Similarly it has a set of guards, that can stop the onbject from
leaving the current state, or entering a next one.

we make use of this to do custom processing on the input when
we arrive to the new state, and stop the state from ever getting
invalid.