URL's now provably correct through state machine.
This is the problem I set out to resolve by rewriting this
using state.js. When isolated from the rest of the application,
I was able to write these tests and ensure that the url's
are always correctly generated.

It took me longer to do since I was carefully capturing
these commits, and writing the commentary.

Still, It has allowed me to reduce the possible corner cases, and
work on just this little piece until it works Right.

next, we rewrite the rest of the application to build
off this new base.